Add new fit function
Add lin_circ_lin fit, which are two linear functions connected by a
circle
segment of a circle to which both lines are tangent, the radius of the
circle
is also a parameter. Add residual plots to PMT data plots. Plot
residuals for
KASCADE fit. Add Grapher file in which the derivation of the
lin_circ_lin is
illustrated.